Made constructors for the FIRKernels.
The FIRFilter constructor was getting messy. It was really a FIRKernel and FIRFilter constructor. In order to test wether it even makes sense having FIRDecimator and FIRInterpolator types when FIRRational should be able to do the same job, I needed to make it easier to construct the kernels explicityly and not have the FIRFilter constructor recogize that 1//2 should be decimator when in fact I want a FIRRational that only does decimation.